fix: cursor postion in trigger
When the code in the trigger executes some code that may affect the cursor position (e.g. a simple #split).

The easiest way to get around this is to use `#delay 0 #split`, but I think the root cause of it is due to the fact that the print mud output doesn't ensure that it must be print in the scroll region.

For code readability reasons, I added a function to do this.

This should close #113